+++ uncertainties/core/affinescalarfunc/base.py
@@ -1,0 +1,468 @@
+
+from __future__ import division  # Many analytical derivatives depend on this
+
+from builtins import str, next, map, zip, range, object
+import math
+from math import sqrt, log, isnan, isinf  # Optimization: no attribute look-up
+import re
+import sys
+try:
+    from math import isinfinite  # !! Python 3.2+
+except ImportError:
+    def isinfinite(x):
+        return isinf(x) or isnan(x)
+
+import copy
+import warnings
+import itertools
+import inspect
+import numbers
+import collections
+
+# The following restricts the local function getargspec() to the common
+# features of inspect.getargspec() and inspect.getfullargspec():
+if sys.version_info < (3,):  # !! Could be removed when moving to Python 3 only
+    from inspect import getargspec
+else:
+    from inspect import getfullargspec as getargspec
+    
+from . formatting import UFloatFormatting
+from .. util import (set_doc, covariance_matrix, partial_derivative,
+NumericalDerivatives, IndexableIter, basestring,
+)
+from .. compat import (deprecation, FLOAT_LIKE_TYPES, CONSTANT_TYPES, CallableStdDev,
+)
+
+
+class LinearCombination(object):
+    """
+    Linear combination of Variable differentials.
+
+    The linear_combo attribute can change formally, but its value
+    always remains the same. Typically, the linear combination can
+    thus be expanded.
+
+    The expanded form of linear_combo is a mapping from Variables to
+    the coefficient of their differential.
+    """
+
+    # ! Invariant: linear_combo is represented internally exactly as
+    # the linear_combo argument to __init__():
+    __slots__ = "linear_combo"
+
+    def __init__(self, linear_combo):
+        """
+        linear_combo can be modified by the object, during its
+        lifetime. This allows the object to change its internal
+        representation over time (for instance by expanding the linear
+        combination and replacing the original expression with the
+        expanded one).
+
+        linear_combo -- if linear_combo is a dict, then it represents
+        an expanded linear combination and must map Variables to the
+        coefficient of their differential. Otherwise, it should be a
+        list of (coefficient, LinearCombination) pairs (that
+        represents a linear combination expression).
+        """
+
+        self.linear_combo = linear_combo
+
+    def __bool__(self):
+        """
+        Return True only if the linear combination is non-empty, i.e. if
+        the linear combination contains any term.
+        """
+        return bool(self.linear_combo)
+
+    def expanded(self):
+        """
+        Return True if and only if the linear combination is expanded.
+        """
+        return isinstance(self.linear_combo, dict)
+
+    def expand(self):
+        """
+        Expand the linear combination.
+
+        The expansion is a collections.defaultdict(float).
+
+        This should only be called if the linear combination is not
+        yet expanded.
+        """
+
+        # The derivatives are built progressively by expanding each
+        # term of the linear combination until there is no linear
+        # combination to be expanded.
+
+        # Final derivatives, constructed progressively:
+        derivatives = collections.defaultdict(float)
+
+        while self.linear_combo:  # The list of terms is emptied progressively
+
+            # One of the terms is expanded or, if no expansion is
+            # needed, simply added to the existing derivatives.
+            #
+            # Optimization note: since Python's operations are
+            # left-associative, a long sum of Variables can be built
+            # such that the last term is essentially a Variable (and
+            # not a NestedLinearCombination): popping from the
+            # remaining terms allows this term to be quickly put in
+            # the final result, which limits the number of terms
+            # remaining (and whose size can temporarily grow):
+            (main_factor, main_expr) = self.linear_combo.pop()
+
+            # print "MAINS", main_factor, main_expr
+
+            if main_expr.expanded():
+                for (var, factor) in main_expr.linear_combo.items():
+                    derivatives[var] += main_factor*factor
+
+            else:  # Non-expanded form
+                for (factor, expr) in main_expr.linear_combo:
+                    # The main_factor is applied to expr:
+                    self.linear_combo.append((main_factor*factor, expr))
+
+            # print "DERIV", derivatives
+
+        self.linear_combo = derivatives
+
+    def __getstate__(self):
+        # Not false, otherwise __setstate__() will not be called:
+        return (self.linear_combo,)
+
+    def __setstate__(self, state):
+        (self.linear_combo,) = state
+
+class AffineScalarFuncBase(object):
+    """
+    Affine functions that support basic mathematical operations
+    (addition, etc.).  Such functions can for instance be used for
+    representing the local (linear) behavior of any function.
+
+    This class can also be used to represent constants.
+
+    The variables of affine scalar functions are Variable objects.
+
+    AffineScalarFunc objects include facilities for calculating the
+    'error' on the function, from the uncertainties on its variables.
+
+    Main attributes and methods:
+
+    - nominal_value, std_dev: value at the origin / nominal value, and
+      standard deviation.  The standard deviation can be NaN or infinity.
+
+    - n, s: abbreviations for nominal_value and std_dev.
+
+    - error_components(): error_components()[x] is the error due to
+      Variable x.
+
+    - derivatives: derivatives[x] is the (value of the) derivative
+      with respect to Variable x.  This attribute is a Derivatives
+      dictionary whose keys are the Variable objects on which the
+      function depends. The values are the numerical values of the
+      derivatives.
+
+      All the Variable objects on which the function depends are in
+      'derivatives'.
+
+    - std_score(x): position of number x with respect to the
+      nominal value, in units of the standard deviation.
+    """
+
+    # To save memory in large arrays:
+    __slots__ = ('_nominal_value', '_linear_part')
+
+    # !! Fix for mean() in NumPy 1.8.0:
+    class dtype(object):
+        type = staticmethod(lambda value: value)
+
+    #! The code could be modified in order to accommodate for non-float
+    # nominal values.  This could for instance be done through
+    # the operator module: instead of delegating operations to
+    # float.__*__ operations, they could be delegated to
+    # operator.__*__ functions (while taking care of properly handling
+    # reverse operations: __radd__, etc.).
+
+    def __init__(self, nominal_value, linear_part):
+        """
+        nominal_value -- value of the function when the linear part is
+        zero.
+
+        linear_part -- LinearCombination that describes the linear
+        part of the AffineScalarFunc.
+        """
+
+        # ! A technical consistency requirement is that the
+        # linear_part can be nested inside a NestedLinearCombination
+        # (because this is how functions on AffineScalarFunc calculate
+        # their result: by constructing nested expressions for them).
+
+        # Defines the value at the origin:
+
+        # Only float-like values are handled.  One reason is that it
+        # does not make sense for a scalar function to be affine to
+        # not yield float values.  Another reason is that it would not
+        # make sense to have a complex nominal value, here (it would
+        # not be handled correctly at all): converting to float should
+        # be possible.
+
+        self._nominal_value = float(nominal_value)
+
+        # In order to have a linear execution time for long sums, the
+        # _linear_part is generally left as is (otherwise, each
+        # successive term would expand to a linearly growing sum of
+        # terms: efficiently handling such terms [so, without copies]
+        # is not obvious, when the algorithm should work for all
+        # functions beyond sums).
+        self._linear_part = linear_part
+
+    # The following prevents the 'nominal_value' attribute from being
+    # modified by the user:
+    @property
+    def nominal_value(self):
+        "Nominal value of the random number."
+        return self._nominal_value
+
+    # Abbreviation (for formulas, etc.):
+    n = nominal_value
+
+    ############################################################
+
+    # Making derivatives a property gives the user a clean syntax,
+    # which is consistent with derivatives becoming a dictionary.
+    @property
+    def derivatives(self):
+        """
+        Return a mapping from each Variable object on which the function
+        (self) depends to the value of the derivative with respect to
+        that variable.
+
+        This mapping should not be modified.
+
+        Derivative values are always floats.
+
+        This mapping is cached, for subsequent calls.
+        """
+
+        if not self._linear_part.expanded():
+            self._linear_part.expand()
+            # Attempts to get the contribution of a variable that the
+            # function does not depend on raise a KeyError:
+            self._linear_part.linear_combo.default_factory = None
+
+        return self._linear_part.linear_combo
+
+    # Uncertainties handling:
+
+    def error_components(self):
+        """
+        Individual components of the standard deviation of the affine
+        function (in absolute value), returned as a dictionary with
+        Variable objects as keys. The returned variables are the
+        independent variables that the affine function depends on.
+
+        This method assumes that the derivatives contained in the
+        object take scalar values (and are not a tuple, like what
+        math.frexp() returns, for instance).
+        """
+
+        # Calculation of the variance:
+        error_components = {}
+
+        for (variable, derivative) in self.derivatives.items():
+
+            # print "TYPE", type(variable), type(derivative)
+
+            # Individual standard error due to variable:
+
+            # 0 is returned even for a NaN derivative (in this case no
+            # multiplication by the derivative is performed): an exact
+            # variable obviously leads to no uncertainty in the
+            # functions that depend on it.
+            if variable._std_dev == 0:
+                # !!! Shouldn't the errors always be floats, as a
+                # convention of this module?
+                error_components[variable] = 0
+            else:
+                error_components[variable] = abs(derivative*variable._std_dev)
+
+        return error_components
+
+    @property
+    def std_dev(self):
+        """
+        Standard deviation of the affine function.
+
+        This method assumes that the function returns scalar results.
+
+        This returned standard deviation depends on the current
+        standard deviations [std_dev] of the variables (Variable
+        objects) involved.
+        """
+        #! It would be possible to not allow the user to update the
+        #std dev of Variable objects, in which case AffineScalarFunc
+        #objects could have a pre-calculated or, better, cached
+        #std_dev value (in fact, many intermediate AffineScalarFunc do
+        #not need to have their std_dev calculated: only the final
+        #AffineScalarFunc returned to the user does).
+        return CallableStdDev(sqrt(sum(
+            delta**2 for delta in self.error_components().values())))
+
+    # Abbreviation (for formulas, etc.):
+    s = std_dev
+
+    def __repr__(self):
+        # Not putting spaces around "+/-" helps with arrays of
+        # Variable, as each value with an uncertainty is a
+        # block of signs (otherwise, the standard deviation can be
+        # mistaken for another element of the array).
+
+        std_dev = self.std_dev  # Optimization, since std_dev is calculated
+
+        # A zero standard deviation is printed because otherwise,
+        # ufloat_fromstr() does not correctly parse back the value
+        # ("1.23" is interpreted as "1.23(1)"):
+
+        if std_dev:
+            std_dev_str = repr(std_dev)
+        else:
+            std_dev_str = '0'
+
+        return "%r+/-%s" % (self.nominal_value, std_dev_str)
+
+    def __str__(self):
+        # An empty format string and str() usually return the same
+        # string
+        # (http://docs.python.org/2/library/string.html#format-specification-mini-language):
+        return self.format('')
+
+    def std_score(self, value):
+        """
+        Return 'value' - nominal value, in units of the standard
+        deviation.
+
+        Raises a ValueError exception if the standard deviation is zero.
+        """
+        try:
+            # The ._nominal_value is a float: there is no integer division,
+            # here:
+            return (value - self._nominal_value) / self.std_dev
+        except ZeroDivisionError:
+            raise ValueError("The standard deviation is zero:"
+                             " undefined result")
+
+    def __deepcopy__(self, memo):
+        """
+        Hook for the standard copy module.
+
+        The returned AffineScalarFunc is a completely fresh copy,
+        which is fully independent of any variable defined so far.
+        New variables are specially created for the returned
+        AffineScalarFunc object.
+        """
+        return self.__class__(self._nominal_value,
+                                copy.deepcopy(self._linear_part))
+
+    def __getstate__(self):
+        """
+        Hook for the pickle module.
+
+        The slot attributes of the parent classes are returned, as
+        well as those of the __dict__ attribute of the object (if
+        any).
+        """
+
+        # In general (case where this class is subclassed), data
+        # attribute are stored in two places: possibly in __dict_, and
+        # in slots. Data from both locations is returned by this
+        # method.
+
+        all_attrs = {}
+
+        # Support for subclasses that do not use __slots__ (except
+        # through inheritance): instances have a __dict__
+        # attribute. The keys in this __dict__ are shadowed by the
+        # slot attribute names (reference:
+        # http://stackoverflow.com/questions/15139067/attribute-access-in-python-first-slots-then-dict/15139208#15139208).
+        # The method below not only preserves this behavior, but also
+        # saves the full contents of __dict__. This is robust:
+        # unpickling gives back the original __dict__ even if __dict__
+        # contains keys that are shadowed by slot names:
+
+        try:
+            all_attrs['__dict__'] = self.__dict__
+        except AttributeError:
+            pass
+
+        # All the slot attributes are gathered.
+
+        # Classes that do not define __slots__ have the __slots__ of
+        # one of their parents (the first parent with their own
+        # __slots__ in MRO). This is why the slot names are first
+        # gathered (with repetitions removed, in general), and their
+        # values obtained later.
+
+        all_slots = set()
+
+        for cls in type(self).mro():
+
+            # In the diamond inheritance pattern, some parent classes
+            # may not have __slots__:
+            slot_names = getattr(cls, '__slots__', ())
+
+            # Slot names can be given in various forms (string,
+            # sequence, iterable):
+            if isinstance(slot_names, basestring):
+                all_slots.add(slot_names)  # Single name
+            else:
+                all_slots.update(slot_names)
+
+        # The slot values are stored:
+        for name in all_slots:
+            try:
+                # !! It might happen that '__dict__' is itself a slot
+                # name. In this case, its value is saved
+                # again. Alternatively, the loop could be done on
+                # all_slots - {'__dict__'}:
+                all_attrs[name] = getattr(self, name)
+            except AttributeError:
+                pass  # Undefined slot attribute
+
+        return all_attrs
+
+    def __setstate__(self, data_dict):
+        """
+        Hook for the pickle module.
+        """
+        for (name, value) in data_dict.items():
+            # Contrary to the default __setstate__(), this does not
+            # necessarily save to the instance dictionary (because the
+            # instance might contain slots):
+            setattr(self, name, value)
+
+    @classmethod
+    def _to_affine_scalar(cls, x):
+        """
+        Transforms x into a constant affine scalar function
+        (AffineScalarFunc), unless it is already an AffineScalarFunc (in
+        which case x is returned unchanged).
+
+        Raises an exception unless x belongs to some specific classes of
+        objects that are known not to depend on AffineScalarFunc objects
+        (which then cannot be considered as constants).
+        """
+
+        if hasattr(x, 'nominal_value') and hasattr(x, 'std_dev'):
+            return x
+
+        if isinstance(x, CONSTANT_TYPES):
+            # No variable => no derivative:
+            return cls(x, LinearCombination({}))
+
+        # Case of lists, etc.
+        raise NotUpcast("%s cannot be converted to a number with"
+                        " uncertainty" % type(x))
+
+
+class NotUpcast(Exception):
+    'Raised when an object cannot be converted to a number with uncertainty'
