+++ uncertainties/core/parsing.py
@@ -1,5 +1,5 @@
 import re
-from . formatting import nrmlze_superscript
+from .affinescalarfunc.formatting import nrmlze_superscript
 
 ###############################################################################
 # Parsing of values with uncertainties:
